refactor(payment-gateways): split interface definitions
